Separate trace event decoders from format information
Previously, trace event decoders and format information were bundled
together, which meant that there could not be two different decoders
for the same tracepoint, and also that if two or more decoders for
the same tracepoint were added, removing one would remove them all.

Format information is now stored in a map like the old decoder information
was, but decoder functions are kept separate as part of the EventMonitor.
Format information is cached permanently for tracepoints, but for kprobes
and uprobes will be removed when the kprobe or uprobe is removed.
